# configuration-service/swagger.yaml
swagger: '2.0'
info:
  title: Configuration Service API
  version: 0.1.0
basePath: /v1
produces:
- application/json
consumes:
  - application/cloudevents+json
  - application/json
schemes:
  - http
definitions:
  Error:
    type: object
    required:
      - message
    properties: 
      code:
        type: integer
        format: int64
        description: Error code
      message:
        type: string
        description: Error message
  Version:
    type: object
    properties:
      version:
        type: string
        description: Version identifier
...
  ExpandedStage:
    type: object
    properties:
      stageName:
        type: string
        description: Stage name
      lastEventContext:
        $ref: '#/definitions/EventContext'
      services:
        type: array
        items:
          $ref: '#/definitions/ExpandedService'
  ExpandedService:
    type: object
    properties:
      serviceName:
        type: string
        description: Service name
      deployedImage:
        type: string
        description: Currently deployed image
      creationDate:
        type: string
        description: Creation date of the service
      lastEventTypes:
        type: object
        additionalProperties:
          $ref: '#/definitions/EventContext'
...
  EventContext:
    type: object
    properties:
      keptnContext:
        type: string
        description: Keptn Context ID of the event
      eventId:
        type: string
        description: ID of the event
      time:
        type: string
        description: Time of the event
...
  Stage:
    type: object
    properties:
      stageName:
        type: string
        description: Stage name
      services:
        type: array
        items:
          $ref: '#/definitions/Service'
  Stages:
    type: object
    properties:
      nextPageKey: 
        type: string
        description: Pointer to next page, base64 encoded
      totalCount: 
        type: number
        description: Total number of stages
      pageSize: 
        type: number
        description: Size of returned page
      stages:
        type: array
        items:
          $ref: '#/definitions/ExpandedStage'
  Service:
    type: object
    properties:
      serviceName:
        type: string
        description: Service name
...
  Resource:
    type: object
    required:
      - resourceURI
    properties:
      resourceURI:
        type: string
        description: Resource URI
      resourceContent:
        type: string
        # format: byte
        description: Resource content
  Resources:
    type: object
    properties:
      nextPageKey: 
        type: string
        description: Pointer to next page, base64 encoded
      totalCount: 
        type: number
        description: Total number of resources
      pageSize: 
        type: number
        description: Size of returned page
      resources:
        type: array
        items:
          $ref: '#/definitions/Resource'
...
parameters:
...
  projectName:
    in: path
    name: projectName
    required: true
    type: string
    description: Name of the project
  stage:
    in: body
    name: stage
    description: Stage entity
    schema:
      $ref: '#/definitions/Stage'
  stageName:
    in: path
    name: stageName
    required: true
    type: string
    description: Name of the stage
...
  serviceName:
    in: path
    name: serviceName
    required: true
    type: string
    description: Name of the service
  resourceURI:
    in: path
    name: resourceURI
    required: true
    type: string
    description: Resource URI
  disableUpstreamSync:
    in: query
    name: disableUpstreamSync
    required: false
    default: false
    type: boolean
    description: Disable sync of upstream repo before reading content
  resource:
    in: body
    name: resource
    description: Resource
    schema:
      $ref: '#/definitions/Resource'
  resources:
    in: body
    name: resources
    description: List of resources
    schema:
      properties:
        resources:
          type: array
          items:
            $ref: '#/definitions/Resource'    
  pageSize:
    in: query
    name: pageSize
    type: integer
    minimum: 1
    maximum: 50
    default: 20
    description: The number of items to return
  nextPageKey:
    in: query
    name: nextPageKey
    type: string
    description: Pointer to the next set of items
paths:
...
  '/project/{projectName}/resource':
    parameters:
      - $ref: '#/parameters/projectName'
    post:
      tags:
        - Project Resource
      summary: Create list of new resources for the project
      parameters:
        - $ref: '#/parameters/resources'        
      responses:
        '201':
          description: Success. Project resource has been created. The version of the new configuration is returned.
          schema:
            $ref: '#/definitions/Version'
        '400':
          description: Failed. Project resource could not be created.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error' 
    get:
      tags:
        - Project Resource
      parameters:
        - $ref: '#/parameters/pageSize'
        - $ref: '#/parameters/nextPageKey'
        - $ref: '#/parameters/disableUpstreamSync'
      summary: Get list of project resources
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Resources'
        '404':
          description: Failed. Containing project could not be found.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'  
    put:
      tags:
        - Project Resource
      parameters:
        - $ref: '#/parameters/resources'
      summary: Update list of project resources
      responses:
        '201':
          description: Success. Project resources have been updated. The version of the new configuration is returned.
          schema:
            $ref: '#/definitions/Version'
        '400':
          description: Failed. Project resources could not be updated.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
  '/project/{projectName}/resource/{resourceURI}':
    parameters:
      - $ref: '#/parameters/projectName'
      - $ref: '#/parameters/resourceURI'
      - $ref: '#/parameters/disableUpstreamSync'
    get:
      tags:
        - Project Resource
      summary: Get the specified resource
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Resource'
        '404':
          description: Failed. Project resource could not be found.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'            
    put:
      tags:
        - Project Resource
      summary: Update the specified resource
      parameters:
        - $ref: '#/parameters/resource' 
      responses:
        '201':
          description: Success. Project resource has been updated. The version of the new configuration is returned.
          schema:
            $ref: '#/definitions/Version'
        '400':
          description: Failed. Project resource could not be updated.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
    delete:
      tags:
        - Project Resource
      summary: Delete the specified resource
      responses:
        '204':
          description: Success. Project resource has been deleted. Response does not have a body.
        '400':
          description: Failed. Project resource could not be deleted.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
  '/project/{projectName}/stage':
    parameters:
      - $ref: '#/parameters/projectName'
    post:
      tags:
        - Stage
      deprecated: true
      summary: 'INTERNAL Endpoint: Create a new stage by stage name'
      parameters:
        - $ref: '#/parameters/stage'
      responses:
        '204':
          description: Success. Stage has been created. Response does not have a body.
        '400':
          description: Failed. Stage could not be created.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
    get:
      tags:
        - Stage
      parameters:
        - $ref: '#/parameters/pageSize'
        - $ref: '#/parameters/nextPageKey'
        - $ref: '#/parameters/disableUpstreamSync'
      summary: Get list of stages
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Stages'
        '404':
          description: Failed. Containing project could not be found.
          schema:
            $ref: '#/definitions/Error' 
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
  '/project/{projectName}/stage/{stageName}':
    parameters:
      - $ref: '#/parameters/projectName'
      - $ref: '#/parameters/stageName'
    get:
      tags:
        - Stage
      summary: Get the specified stage
      parameters:
        - $ref: '#/parameters/disableUpstreamSync'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ExpandedStage'
        '404':
          description: Failed. Project resource could not be found.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error' 
    put:
      tags:
        - Stage
      deprecated: true
      summary: 'INTERNAL Endpoint: Update the specified stage'
      parameters:
        - $ref: '#/parameters/stage'
      responses:
        '204':
          description: Success. Stage has been updated. Response does not have a body.
        '400':
          description: Failed. Stage could not be updated.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
    delete:
      tags:
        - Stage
      deprecated: true
      summary: 'INTERNAL Endpoint: Delete the specified stage'
      responses:
        '204':
          description: Success. Stage has been deleted. Response does not have a body.
        '400':
          description: Failed. Stage could not be deleted.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
  '/project/{projectName}/stage/{stageName}/resource':
    parameters:
      - $ref: '#/parameters/projectName'
      - $ref: '#/parameters/stageName'
    post:
      tags:
        - Stage Resource
      summary: Create list of new resources for the stage
      parameters:
        - $ref: '#/parameters/resources'   
      responses:
        '201':
          description: Success. Stage resource has been created. The version of the new configuration is returned.
          schema:
            $ref: '#/definitions/Version'
        '400':
          description: Failed. Stage resource could not be created.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
    get:
      tags:
        - Stage Resource
      parameters:
        - $ref: '#/parameters/pageSize'
        - $ref: '#/parameters/nextPageKey'
        - $ref: '#/parameters/disableUpstreamSync'
      summary: Get list of stage resources
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Resources'
        '404':
          description: Failed. Stage could not be found.
          schema:
            $ref: '#/definitions/Error' 
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
    put:
      tags:
        - Stage Resource
      parameters:
        - $ref: '#/parameters/resources'
      summary: Update list of stage resources
      responses:
        '201':
          description: Success. Stage resources have been updated. The version of the new configuration is returned.
          schema:
            $ref: '#/definitions/Version'
        '400':
          description: Failed. Stage resources could not be updated.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
  '/project/{projectName}/stage/{stageName}/resource/{resourceURI}':
    parameters:
      - $ref: '#/parameters/projectName'
      - $ref: '#/parameters/stageName'
      - $ref: '#/parameters/resourceURI'
    get:
      tags:
        - Stage Resource
      summary: Get the specified resource
      parameters:
        - $ref: '#/parameters/disableUpstreamSync'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Resource'
        '404':
          description: Failed. Stage resource could not be found.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error' 
    put:
      tags:
        - Stage Resource
      summary: Update the specified resource
      parameters:
        - $ref: '#/parameters/resource'
      responses:
        '201':
          description: Success. Stage resource has been updated. The version of the new configuration is returned.
          schema:
            $ref: '#/definitions/Version'
        '400':
          description: Failed. Stage resource could not be updated.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
    delete:
      tags:
        - Stage Resource
      summary: Delete the specified resource
      responses:
        '204':
          description: Success. Stage resource has been deleted. Response does not have a body.
        '400':
          description: Failed. Stage resource could not be deleted.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
...
  '/project/{projectName}/stage/{stageName}/service/{serviceName}/resource':
    parameters:
      - $ref: '#/parameters/projectName'
      - $ref: '#/parameters/stageName'
      - $ref: '#/parameters/serviceName'
    post:
      tags:
        - Service Resource
      summary: Create list of new resources for the service
      parameters:
        - $ref: '#/parameters/resources'   
      responses:
        '201':
          description: Success. Service resource has been created. The version of the new configuration is returned.
          schema:
            $ref: '#/definitions/Version'
        '400':
          description: Failed. Service resource could not be created.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
    get:
      tags:
        - Service Resource
      parameters:
        - $ref: '#/parameters/pageSize'
        - $ref: '#/parameters/nextPageKey'
        - $ref: '#/parameters/disableUpstreamSync'
      summary: Get list of service resources
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Resources'
        '404':
          description: Failed. Containing Service could not be found.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error' 
    put:
      tags:
        - Service Resource
      parameters:
        - $ref: '#/parameters/resources'
      summary: Update list of service resources
      responses:
        '201':
          description: Success. Service resources have been updated. The version of the new configuration is returned.
          schema:
            $ref: '#/definitions/Version'
        '400':
          description: Failed. Service resources could not be updated.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
  '/project/{projectName}/stage/{stageName}/service/{serviceName}/resource/{resourceURI}':
    parameters:
      - $ref: '#/parameters/projectName'
      - $ref: '#/parameters/stageName'
      - $ref: '#/parameters/serviceName'
      - $ref: '#/parameters/resourceURI'
    get:
      tags:
        - Service Resource
      summary: Get the specified resource
      parameters:
        - $ref: '#/parameters/disableUpstreamSync'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Resource'
        '404':
          description: Failed. Service resource could not be found.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error' 
    put:
      tags:
        - Service Resource
      summary: Update the specified resource
      parameters:
        - $ref: '#/parameters/resource'
      responses:
        '201':
          description: Success. Service resource has been updated. The version of the new configuration is returned.
          schema:
            $ref: '#/definitions/Version'
        '400':
          description: Failed. Service resource could not be updated.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
    delete:
      tags:
        - Service Resource
      summary: Delete the specified resource
      responses:
        '204':
          description: Success. Service resource has been deleted. Response does not have a body.
        '400':
          description: Failed. Service resource could not be deleted.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
  '/project/{projectName}/service/{serviceName}/resource':
    parameters:
      - $ref: '#/parameters/projectName'
      - $ref: '#/parameters/serviceName'
    post:
      tags:
        - Service Default Resource 
      summary: Create list of default resources for the service used in all stages
      parameters:
        - $ref: '#/parameters/resources' 
      responses:
        '201':
          description: Success. Service default resource has been created. The version of the new configuration is returned.
          schema:
            $ref: '#/definitions/Version'
        '400':
          description: Failed. Service default resource could not be created.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
    get:
      tags:
        - Service Default Resource 
      parameters:
        - $ref: '#/parameters/pageSize'
        - $ref: '#/parameters/nextPageKey'
        - $ref: '#/parameters/disableUpstreamSync'
      summary: Get list of default resources for the service used in all stages
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Resources'
        '404':
          description: Failed. Containing service could not be found.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error' 
    put:
      tags:
        - Service Default Resource
      parameters:
        - $ref: '#/parameters/resources'
      summary: Update list of service default resources
      responses:
        '201':
          description: Success. Service default resources have been updated. The version of the new configuration is returned.
          schema:
            $ref: '#/definitions/Version'
        '400':
          description: Failed. Service default resources could not be updated.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
  '/project/{projectName}/service/{serviceName}/resource/{resourceURI}':
    parameters:
      - $ref: '#/parameters/projectName'
      - $ref: '#/parameters/serviceName'
      - $ref: '#/parameters/resourceURI'
    get:
      tags:
        - Service Default Resource
      summary: Get the specified default resource for the service
      parameters:
        - $ref: '#/parameters/disableUpstreamSync'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Resource'
        '404':
          description: Failed. Default resource could not be found.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error' 
    put:
      tags:
        - Service Default Resource
      summary: Update the specified default resource for the service
      responses:
        '201':
          description: Success. Service default resource has been updated. The version of the new configuration is returned.
          schema:
            $ref: '#/definitions/Version'
        '400':
          description: Failed. Service default resource could not be updated.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
    delete:
      tags:
        - Service Default Resource
      summary: Delete the specified default resource for the service
      responses:
        '204':
          description: Success. Service default resource has been deleted. Response does not have a body.
        '400':
          description: Failed. Service default resource could not be deleted.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
...